+++ lib/chef/knife/container_docker_build.rb
@@ -15,13 +15,12 @@
 # limitations under the License.
 #
 
-require 'chef/knife'
-require 'chef/mixin/shell_out'
+require 'chef/knife/container_docker_base'
 
 class Chef
   class Knife
     class ContainerDockerBuild < Knife
-      include Chef::Mixin::ShellOut
+      include Knife::ContainerDockerBase
 
       deps do
         # These two are needed for cleanup
@@ -29,43 +28,55 @@ class Chef
         require 'chef/api_client'
       end
 
-      banner "knife container docker build REPO/NAME [options]"
+      banner 'knife container docker build REPO/NAME [options]'
 
       option :run_berks,
-        :long => "--[no-]berks",
-        :description => "Run Berkshelf",
-        :default => true,
-        :boolean => true
+        long:         '--[no-]berks',
+        description:  'Run Berkshelf',
+        default:      true,
+        boolean:      true
 
       option :berks_config,
-        :long => "--berks-config CONFIG",
-        :description => "Use the specified Berkshelf configuration"
+        long:         '--berks-config CONFIG',
+        description:  'Use the specified Berkshelf configuration'
 
       option :cleanup,
-        :long => "--[no-]cleanup",
-        :description => "Cleanup Chef and Docker artifacts",
-        :default => true,
-        :boolean => true
+        long:         '--[no-]cleanup',
+        description:  'Cleanup Chef and Docker artifacts',
+        default:      true,
+        boolean:      true
+
+      option :secure_dir,
+        long:         '--secure-dir DIR',
+        description:  'Path to a local repository that contains Chef credentials.'
 
       option :force_build,
-        :long => "--force",
-        :description => "Force the Docker image build",
-        :boolean => true
+        long:         '--force',
+        description:  'Force the Docker image build',
+        boolean:      true
+
+      option :tags,
+        long:         '--tags TAG[,TAG]',
+        description:  'Comma separated list of tags you wish to apply to the image.',
+        default:      ['latest'],
+        proc:         proc { |o| o.split(/[\s,]+/) }
 
       option :dockerfiles_path,
-        :short => "-d PATH",
-        :long => "--dockerfiles-path PATH",
-        :description => "Path to the directory where Docker contexts are kept",
-        :proc => Proc.new { |d| Chef::Config[:knife][:dockerfiles_path] = d }
+        short:        '-d PATH',
+        long:         '--dockerfiles-path PATH',
+        description:  'Path to the directory where Docker contexts are kept',
+        proc:         proc { |d| Chef::Config[:knife][:dockerfiles_path] = d }
 
       #
       # Run the plugin
       #
       def run
-        read_and_validate_params
+        validate
         setup_config_defaults
         run_berks if config[:run_berks]
-        build_image
+        backup_secure unless config[:secure_dir].nil?
+        build_docker_image
+        restore_secure unless config[:secure_dir].nil?
         cleanup_artifacts if config[:cleanup]
       end
 
@@ -73,23 +84,44 @@ class Chef
       # Reads the input parameters and validates them.
       # Will exit if it encounters an error
       #
-      def read_and_validate_params
+      def validate
         if @name_args.length < 1
           show_usage
-          ui.fatal("You must specify a Dockerfile name")
+          ui.fatal('You must specify a Dockerfile name')
           exit 1
         end
 
-        # if berkshelf isn't installed, set run_berks to false
-        if config[:run_berks]
-          ver = shell_out("berks -v")
-          config[:run_berks] = ver.stdout.match(/\d+\.\d+\.\d+/) ? true : false
+        unless valid_dockerfile_name?(@name_args[0])
+          show_usage
+          ui.fatal('Your Dockerfile name cannot include a protocol or a tag.')
+          exit 1
+        end
 
-          if config[:berks_config]
-            unless File.exists?(config[:berks_config])
-              ui.fatal("No Berksfile configuration found at #{config[:berks_config]}")
-              exit 1
-            end
+        # if secure_dir doesn't exist or is missing files, exit
+        if config[:secure_dir]
+          case
+          when !File.directory?(config[:secure_dir])
+            ui.fatal("SECURE_DIRECTORY: The directory #{config[:secure_dir]}" \
+              " does not exist.")
+            exit 1
+          when !File.exist?(File.join(config[:secure_dir], 'validation.pem')) &&
+            !File.exist?(File.join(config[:secure_dir], 'client.pem'))
+            ui.fatal("SECURE_DIRECTORY: Can not find validation.pem or client.pem" \
+              " in #{config[:secure_dir]}.")
+            exit 1
+          end
+        end
+
+        # if berkshelf isn't installed, set run_berks to false
+        unless berks_installed?
+          ui.warn('The berks executable could not be found. Resolving the Berksfile will be skipped.')
+          config[:run_berks] = false
+        end
+
+        if config[:berks_config]
+          unless File.exist?(config[:berks_config])
+            ui.fatal("No Berksfile configuration found at #{config[:berks_config]}")
+            exit 1
           end
         end
       end
@@ -98,18 +130,18 @@ class Chef
       # Set defaults for configuration values
       #
       def setup_config_defaults
-        Chef::Config[:knife][:dockerfiles_path] ||= File.join(Chef::Config[:chef_repo_path], "dockerfiles")
+        Chef::Config[:knife][:dockerfiles_path] ||= File.join(Chef::Config[:chef_repo_path], 'dockerfiles')
         config[:dockerfiles_path] = Chef::Config[:knife][:dockerfiles_path]
 
         # Determine if we are running local or server mode
         case
-        when File.exists?(File.join(config[:dockerfiles_path], @name_args[0], 'chef', 'zero.rb'))
+        when File.exist?(File.join(config[:dockerfiles_path], dockerfile_name, 'chef', 'zero.rb'))
           config[:local_mode] = true
-        when File.exists?(File.join(config[:dockerfiles_path], @name_args[0], 'chef', 'client.rb'))
+        when File.exist?(File.join(config[:dockerfiles_path], dockerfile_name, 'chef', 'client.rb'))
           config[:local_mode] = false
         else
           show_usage
-          ui.fatal("Can not find a Chef configuration file in #{config[:dockerfiles_path]}/#{@name_args[0]}/chef")
+          ui.fatal("Can not find a Chef configuration file in #{config[:dockerfiles_path]}/#{dockerfile_name}/chef")
           exit 1
         end
       end
@@ -118,10 +150,10 @@ class Chef
       # Execute berkshelf locally
       #
       def run_berks
-        if File.exists?(File.join(docker_context, "Berksfile"))
-          if File.exists?(File.join(chef_repo, "zero.rb"))
+        if File.exist?(File.join(docker_context, 'Berksfile'))
+          if File.exist?(File.join(chef_repo, 'zero.rb'))
             run_berks_vendor
-          elsif File.exists?(File.join(chef_repo, "client.rb"))
+          elsif File.exist?(File.join(chef_repo, 'client.rb'))
             run_berks_upload
           end
         end
@@ -133,26 +165,27 @@ class Chef
       # @returns [TrueClass, FalseClass]
       #
       def berksfile_exists?
-        File.exists?(File.join(docker_context, "Berksfile"))
+        File.exist?(File.join(docker_context, 'Berksfile'))
       end
+
 
       #
       # Installs all the cookbooks via Berkshelf
       #
       def run_berks_install
-        run_command("berks install")
+        run_command('berks install')
       end
 
       #
       # Vendors all the cookbooks into a directory inside the Docker Context
       #
       def run_berks_vendor
-        if File.exists?(File.join(chef_repo, "cookbooks"))
+        if File.exist?(File.join(chef_repo, 'cookbooks'))
           if config[:force_build]
-            FileUtils.rm_rf(File.join(chef_repo, "cookbooks"))
+            FileUtils.rm_rf(File.join(chef_repo, 'cookbooks'))
           else
             show_usage
-            ui.fatal("A `cookbooks` directory already exists. You must either remove this directory from your dockerfile directory or use the `force` flag")
+            ui.fatal('A `cookbooks` directory already exists. You must either remove this directory from your dockerfile directory or use the `force` flag')
             exit 1
           end
         end
@@ -166,8 +199,8 @@ class Chef
       #
       def run_berks_upload
         run_berks_install
-        berks_upload_cmd = "berks upload"
-        berks_upload_cmd << " --force" if config[:force_build]
+        berks_upload_cmd = 'berks upload'
+        berks_upload_cmd << ' --force' if config[:force_build]
         berks_upload_cmd << " --config=#{File.expand_path(config[:berks_config])}" if config[:berks_config]
         run_command(berks_upload_cmd)
       end
@@ -175,8 +208,38 @@ class Chef
       #
       # Builds the Docker image
       #
-      def build_image
-        run_command(docker_build_command)
+      def build_docker_image
+        image = build_image(docker_context)
+
+        # Apply the name and tags
+        # By default, it will apply the latest tag but this behavior can be
+        # overwritten by excluding the latest tag from the specified list.
+        config[:tags].each do |tag|
+          tag_image(image.id, @name_args[0], tag)
+        end
+      end
+
+      #
+      # Move `secure` folder to `backup_secure` and copy the config[:secure_dir]
+      # to the new `secure` folder.
+      #
+      # Note: The .dockerignore file has a line to ignore backup secure to it
+      #       should not be included in the image.
+      #
+      def backup_secure
+        FileUtils.mv("#{docker_context}/chef/secure",
+          "#{docker_context}/chef/secure_backup")
+        FileUtils.cp_r(config[:secure_dir], "#{docker_context}/chef/secure")
+      end
+
+      #
+      # Delete the temporary secure directory and restore the original from the
+      # backup.
+      #
+      def restore_secure
+        FileUtils.rm_rf("#{docker_context}/chef/secure")
+        FileUtils.mv("#{docker_context}/chef/secure_backup",
+          "#{docker_context}/chef/secure")
       end
 
       #
@@ -184,16 +247,9 @@ class Chef
       #
       def cleanup_artifacts
         unless config[:local_mode]
-          destroy_item(Chef::Node, node_name, "node")
-          destroy_item(Chef::ApiClient, node_name, "client")
+          destroy_item(Chef::Node, node_name, 'node')
+          destroy_item(Chef::ApiClient, node_name, 'client')
         end
-      end
-
-      #
-      # The command to use to build the Docker image
-      #
-      def docker_build_command
-        "docker build -t #{@name_args[0]} #{docker_context}"
       end
 
       #
@@ -214,7 +270,14 @@ class Chef
       # @return [String]
       #
       def docker_context
-        File.join(config[:dockerfiles_path], @name_args[0])
+        File.join(config[:dockerfiles_path], dockerfile_name)
+      end
+
+      #
+      # Returns the encoded Dockerfile name
+      #
+      def dockerfile_name
+        parse_dockerfile_name(@name_args[0])
       end
 
       #
@@ -223,7 +286,7 @@ class Chef
       # @return [String]
       #
       def chef_repo
-        File.join(docker_context, "chef")
+        File.join(docker_context, 'chef')
       end
 
       #
@@ -232,7 +295,7 @@ class Chef
       # @return [String]
       #
       def node_name
-        File.read(File.join(chef_repo, ".node_name")).strip
+        File.read(File.join(chef_repo, '.node_name')).strip
       end
 
       # Extracted from Chef::Knife.delete_object, because it has a