+++ lib/knife-container/generator.rb
@@ -29,6 +29,7 @@ module KnifeContainer
       attr_accessor :cookbook_path
       attr_accessor :role_path
       attr_accessor :node_path
+      attr_accessor :data_bag_path
       attr_accessor :environment_path
       attr_accessor :chef_server_url
       attr_accessor :validation_key
@@ -37,6 +38,7 @@ module KnifeContainer
       attr_accessor :encrypted_data_bag_secret
       attr_accessor :first_boot
       attr_accessor :berksfile
+      attr_accessor :berksfile_source
       attr_accessor :generate_berksfile
       attr_accessor :run_berks
       attr_accessor :force_build
@@ -77,6 +79,7 @@ module KnifeContainer
       delegate_to_app_context :encrypted_data_bag_secret
       delegate_to_app_context :first_boot
       delegate_to_app_context :berksfile
+      delegate_to_app_context :berksfile_source
       delegate_to_app_context :generate_berksfile
       delegate_to_app_context :run_berks
       delegate_to_app_context :force_build